+++ src/main/java/chess/view/OutputView.java
@@ -8,9 +8,9 @@ import chess.domain.piece.position.Rank;
 import chess.domain.piece.type.Bishop;
 import chess.domain.piece.type.King;
 import chess.domain.piece.type.Knight;
-import chess.domain.piece.type.Pawn;
 import chess.domain.piece.type.Queen;
 import chess.domain.piece.type.Rook;
+import chess.domain.piece.type.pawn.Pawn;
 
 import java.util.HashMap;
 import java.util.List;
