+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -6,9 +6,9 @@ import chess.domain.piece.position.PiecePosition;
 import chess.domain.piece.type.Bishop;
 import chess.domain.piece.type.King;
 import chess.domain.piece.type.Knight;
-import chess.domain.piece.type.Pawn;
 import chess.domain.piece.type.Queen;
 import chess.domain.piece.type.Rook;
+import chess.domain.piece.type.pawn.Pawn;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
