+++ src/test/java/chess/domain/piece/move/WhitePawnMoveStrategyTest.java
@@ -2,6 +2,7 @@ package chess.domain.piece.move;
 
 import chess.domain.piece.position.Path;
 import chess.domain.piece.position.PiecePosition;
+import chess.domain.piece.type.pawn.move.WhitePawnMoveStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
